make sure we have all the tracking branches around
Travis clones the repo with `--single-branch` which prevents us from
accessing the branches we need. This commit fetches all the remote
branches, and creates local tracking branches.

For PRs, we need to use the commit that correspond to the PR instead of
the branch on which the PR is based. This is also handled.